Add files via upload
University Management System developed using the C++ programming language. The project aims to efficiently manage both the academic and administrative aspects of a university in a user-friendly manner.
The choice of implementing this University Management System in C++ was a deliberate one, owing to its robust capabilities and versatility in handling complex tasks. By utilizing the C++ programming language, the project ensures a seamless user interface and reliable functionality.
Furthermore, the project incorporates the eSpeak software, an open-source text-to-speech (TTS) engine. This software enhances accessibility and user experience by converting textual information into natural-sounding speech. Through the integration of eSpeak, the system provides auditory feedback and notifications to users, contributing to a more inclusive and engaging environment.